# repo-stat/processor/internal/adapter/repository/gen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package generated_processor_db

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type InboxMessage struct {
	ID        uuid.UUID
	Payload   []byte
	CreatedAt pgtype.Timestamptz
}

type OutboxMessage struct {
	ID        uuid.UUID
	Repo      string
	Owner     string
	Status    string
	CreatedAt pgtype.Timestamptz
}

type Repository struct {
	ID          uuid.UUID
	Fullname    string
	Description string
	Forks       int32
	Stargazers  int32
	CreatedAt   pgtype.Timestamptz
	Status      string
	UpdatedAt   pgtype.Timestamptz
}

type Subscription struct {
	ID        uuid.UUID
	RepoName  string
	OwnerName string
	CreatedAt pgtype.Timestamptz
}